Move to spectator to collect metrics. Default binding is to NoopRegistry. Refer to https://github.com/Netflix/spectator for more details.

+++ priam/src/main/java/com/netflix/priam/google/GoogleEncryptedFileSystem.java
@@ -316,7 +316,7 @@ public class GoogleEncryptedFileSystem implements IBackupFileSystem, GoogleEncry
     }
 
     @Override
-    public int getAWSSlowDownExceptionCounter() {
+    public long getAWSSlowDownExceptionCounter() {
         return 0;
     }
